Add Instagram post ingestion and related models

+++ app/models/comment.py
@@ -1,4 +1,7 @@
-from sqlalchemy import Column, Integer, String, Text, DateTime, Enum as SQLEnum
+from datetime import datetime
+from tokenize import Double
+
+from sqlalchemy import Column, Integer, String, Text, DateTime, Float, Enum as SQLEnum
 from sqlalchemy.sql import func
 from app.core.database import Base
 import enum
@@ -23,10 +26,11 @@ class Comment(Base):
     id_comment_platform = Column(Integer, nullable=True)
     author = Column(String(255), nullable=True)
     content = Column(Text, nullable=False)
-    rating = Column(Integer, nullable=True)
+    rating = Column(Float, nullable=True)
     post_url = Column(String(512), nullable=True)
     created_at = Column(DateTime(timezone=True), server_default=func.now(), nullable=False)
     platform_created_at = Column(DateTime(timezone=True), nullable=True)
 
     def __repr__(self):
         return f"<Comment(id={self.id}, platform={self.platform}, author={self.author})>"
+
